+++ unicorn_binance_trailing_stop_loss/__main__.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 #
-# File: unicorn_binance_trailing_stop_loss/__init__.py
+# File: unicorn_binance_trailing_stop_loss/__main__.py
 #
 # Part of ‘UNICORN Binance Trailing Stop Loss’
 # Project website: https://www.lucit.tech/unicorn-binance-trailing-stop-loss.html
@@ -34,5 +34,6 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS
 # IN THE SOFTWARE.
 
-from unicorn_binance_trailing_stop_loss.manager import BinanceTrailingStopLossManager
+from .cli import main
 
+main()